Henna-Solutions_README.md
Henna Solutions, a software company, is working on a web application project. The company aims to create certain HTML Web Pages for the Authentication flow in the project for their client. Developers and programmers in the project development team need to sync the code to Github to work on code collaboratively.